stash abandond code in branch
This was an attempt to allow the retrieval to run over seasons.
This attempt was abandoned to start in a fresh repository.
New Northampton inversion that can have any start or end date is
in the Assimila repository EOLDAS_family (ADAS folder). This code
is being stashed in a branch  for now.